[serde-generate] split file for runtime tests, create feature "runtime-testing"
